lab7/requirements.txt
websocket-client==0.48.0
cryptography==2.7
py-bcrypt==0.4